refactor: GoogleSignInClient 클래스 분리
