+++ app/src/main/java/com/fintern/ourbudgeting/data/auth/GoogleSignInClient.kt
@@ -1,0 +1,31 @@
+package com.fintern.ourbudgeting.data.auth
+
+import android.content.Context
+import androidx.credentials.CredentialManager
+import androidx.credentials.GetCredentialRequest
+import com.fintern.ourbudgeting.BuildConfig
+import com.google.android.libraries.identity.googleid.GetGoogleIdOption
+import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
+import javax.inject.Inject
+
+class GoogleSignInClient @Inject constructor(
+    private val context: Context,
+    private val credentialManager: CredentialManager
+) {
+    suspend fun getGoogleIdToken(): String {
+        val googleIdOption = GetGoogleIdOption.Builder()
+            .setFilterByAuthorizedAccounts(false)
+            .setServerClientId(BuildConfig.GOOGLE_CLIENT_ID)
+            .build()
+
+        val request = GetCredentialRequest.Builder()
+            .addCredentialOption(googleIdOption)
+            .build()
+
+        val result = credentialManager.getCredential(context, request)
+        val credential = result.credential
+
+        val googleIdTokenCredential = GoogleIdTokenCredential.createFrom(credential.data)
+        return googleIdTokenCredential.idToken
+    }
+}
